--- a/ocamlbuild/display.mli
+++ b/ocamlbuild/display.mli
@@ -31,4 +31,5 @@ val finish : ?how:[`Success|`Error|`Quiet] -> display -> unit
 val event : display -> ?pretend:bool -> string -> string -> Tags.t -> unit
 val display : display -> (out_channel -> unit) -> unit
 val update : display -> unit
+val is_logging : display -> int -> bool
 val dprintf : ?raw:bool -> ?log_level:int -> display -> ('a, Format.formatter, unit) format -> 'a