--- a/ocamlbuild/signatures.mli
+++ b/ocamlbuild/signatures.mli
@@ -312,6 +312,9 @@ module type LOG = sig
       ["@\[<2>"] and ["@\]@."]. *)
   val dprintf : int -> ('a, Format.formatter, unit) format -> 'a
 
+  (** @return whether logging at this level will print anything *)
+  val is_logging : int -> bool
+
   (** Equivalent to calling [dprintf] with a level [< 0]. *)
   val eprintf : ('a, Format.formatter, unit) format -> 'a