See http://caml.inria.fr/mantis/view.php?id=4934 and http://caml.inria.fr/mantis/view.php?id=5756

Timings:
Building already built coq 8.4 with current trunk takes 51s here.
First commit brings it down to 46s, second - 34s, all three - 30s.
All in all 40% speedup with minimal changes.

Explanation:
1. Specializing `Hashtbl` is usual elimination of generic `compare`.
2. `Log.dprintf` wraps log format even if there will be no printing done - this is too expensive and is fixed globally.
3. Still the recursive procedure in `Solver.self` generates too much garbage in `Format.ifprintf` when discarding printing, this is solved with `if loglevel <= X then dprintf X` for selected hot path (5 calls total).