@gasche 2015-07-27
Thanks! The patch is obviously correct, added to my merge list.