@gasche 2015-08-02
Merged in trunk. Thanks for the patch.